(docs): update Magic Patterns description
Hey! Thanks for originally adding us (co-founder of Magic Patterns here). Saw this list naturally in the wild, which was a pleasure.

One key thing about us is that we support custom design systems (bringing your own component library). Are you open to this new description?

Appreciate it — great list! 🔥