Update README.md
Updated the entry for Amazon Q Developer (included the new features released in December). I also added a separate entry for Amazon Q Developer CLI, which is a separate product and has new agentic capabilities. Links were updated as the home page for these is different. Hope this is ok.